勝利の判定を別ファイルに移動
