+++ src/tictactoe/index.ts
@@ -1,32 +1,9 @@
 import { useState } from "react"
 import { range } from "util/range"
 import { Player, Marker, Status } from "model/tictactoe-interface"
+import { calculateWinner } from "model/calculate-winner"
 
 export type History = readonly Marker[][]
-
-export const calculateWinner = (squares: Marker[]): Player | null => {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ]
-  for (const line of lines) {
-    const [a, b, c] = line
-    const A = squares[a]
-    const B = squares[b]
-    const C = squares[c]
-
-    if (A !== "BLANK" && A === B && A === C) {
-      return A
-    }
-  }
-  return null
-}
 
 export type UseTicTacToeProps = {
   initHistory: History
@@ -50,7 +27,7 @@ export const useTicTacToe = ({
   const [history, setHistory] = useState<History>(initHistory)
 
   const squares = history[stepNumber]
-  const winner: Player | null = calculateWinner(squares)
+  const winner = calculateWinner(squares)
 
   const placeMarker = (index: number) => {
     if (squares[index] || winner) {
